+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
-## 1.1.2
-Change user ID to INT value for dataLayer (return NULL id user is not logged in).
+## 1.2.0
+Click tracking extending and improvements to user ID detection.
+
+### Added
+- Click tracking attributes for button block.
+- Click tracking attributes for Yoast SEO breadcrum.
+
+### Changed
+- Do not add wp_userid property to dataLayer if user is not logged in (should return NULL in GTM).
 
 ## 1.1.1
-Change pageview and click events names for dataLayer
+Change pageview and click events names for dataLayer.
 
 ## 1.1.0
 Add support to detect wp_nav_menu items click via dataLayer.